Base Activity added (need replace commit from shared pref save + toolbar color)
