+++ app/src/main/res/values/strings.xml
@@ -5,7 +5,7 @@
     <string name="my_name">Anton Taleckij</string>
     <string name="my_email">anton.taleckij.job@gmail.com</string>
     <string name="hello_message">Hello. Welcome. Glad to see you!</string>
-    <string name="finish_button_text">Next</string>
+    <string name="finish_button_text">Finish</string>
     <string name="app_description">It is the simple launcher app</string>
 
     <string name="light_theme_text">Light theme</string>
